+++ Migrators/AllureExporter/Models/Constants.cs
@@ -2,7 +2,7 @@ namespace AllureExporter.Models;
 
 public class Constants
 {
-    public const int MainSectionId = 0;
+    public const long MainSectionId = 0;
     public const string AllureStatus = "AllureStatus";
     public const string AllureTestLayer = "Test Layer";
     public const string Feature = "Feature";
