+++ Migrators/AllureExporter/Models/SectionInfo.cs
@@ -7,5 +7,5 @@ namespace AllureExporter.Models;
 public class SectionInfo
 {
     public Section MainSection { get; set; }
-    public Dictionary<int, Guid> SectionDictionary { get; set; }
+    public Dictionary<long, Guid> SectionDictionary { get; set; }
 }
