+++ Migrators/AllureExporter/Services/ITestCaseService.cs
@@ -5,5 +5,5 @@ namespace AllureExporter.Services;
 
 public interface ITestCaseService
 {
-    Task<List<TestCase>> ConvertTestCases(int projectId, Dictionary<string, Guid> sharedStepMap, Dictionary<string, Guid> attributes,  SectionInfo sectionInfo);
+    Task<List<TestCase>> ConvertTestCases(long projectId, Dictionary<string, Guid> sharedStepMap, Dictionary<string, Guid> attributes,  SectionInfo sectionInfo);
 }
